finalize page
