+++ force-app/main/default/lwc/edaSettings/edaSettings.html
@@ -38,6 +38,10 @@
                     <template if:true={settingsPageToDisplay.contactInformationSettings}>
                         <c-contact-information-settings></c-contact-information-settings>
                     </template>
+
+                    <template if:true={settingsPageToDisplay.programSettings}>
+                        <c-program-settings></c-program-settings>
+                    </template>
                 </div>
             </template>
         </template>
